Last Changes before we hand over project
Everything up to this point works and has been test on New zealand.
Comments are comprehensive

What has not been done: The big boi algorithm is not automated, but it works (we were reunning it byt hardcording to get the results). Other than that all works perfectlty fine for NZ